+++ packages/worker/src/api/controllers/global/auth.ts
@@ -140,6 +140,7 @@ export const datasourcePreAuth = async (ctx: any, next: any) => {
     {
       provider,
       appId: ctx.query.appId,
+      datasourceId: ctx.query.datasourceId,
     },
     Cookie.DatasourceAuth
   )
